Add solution for Restaurant Customers problem
A new file for the CSES problem 'Restaurant Customers' has been added. The problem is solved using a vector to store the incoming and outgoing times of the customers at the restaurant. For each time point, the number of customers is incremented or decremented accordingly and the maximum number of customers at any given time is calculated and returned as the solution to the problem.