Add new solution for Sum of Two Values problem
Added a new file to solve the Sum of Two Values problem from CSES. The algorithm applies "two sum" problem approach, where we search through a map to find if the remainder between a target sum and the current value exists. If exists, the indices of these two numbers are outputted, otherwise "IMPOSSIBLE" is printed. The map is used for efficient search of the numbers.