Add solution for the "Missing Coin Sum" problem.
Implemented the solution for the "Missing Coin Sum" problem from CSES in a new file named "Missing Coin Sum.cpp". The solution involves reading the input of coin values, sorting them, and iterating through them to find the smallest possible sum of money that cannot be created with these coins. The sum is initially set to 1 and is increased by current coin value if that coin value is lesser or equal to the current sum. The final sum is the result which we print out.