Add 'Playlist.cpp' for implementation of CSES task
This commit introduces 'Playlist.cpp', which is a new file added to implement the problem from the CSES platform. The problem solved in this file is related to finding the longest playlist without repeated songs. The solution uses two points (i and j) and a map to track already played songs. It keeps increasing the set size and adjusting the pointers until it encounters a song previously played.